requirements.txt
# ═══════════════════════════════════════════════════════════════════════
#  Sport Data Solution — Dépendances Python (v3.1.0)
#  Python >= 3.10 requis
# ═══════════════════════════════════════════════════════════════════════

# ─── Data / ETL ─────────────────────────────────────────────────────────
pandas>=2.2.0,<3.0
numpy>=1.26.0,<2.0
openpyxl>=3.1.2,<4.0

# ─── Base de données ────────────────────────────────────────────────────
sqlalchemy>=2.0.25,<3.0
psycopg[binary]>=3.1.17,<4.0     # driver PostgreSQL moderne
python-dotenv>=1.0.0,<2.0

# ─── Géocodage ──────────────────────────────────────────────────────────
googlemaps>=4.10.0,<5.0          # API Google Maps (primaire)
geopy>=2.4.0,<3.0                # Nominatim (fallback)

# ─── Data Quality ───────────────────────────────────────────────────────
great-expectations>=0.18.0,<1.0

# ─── Utilitaires ────────────────────────────────────────────────────────
requests>=2.31.0,<3.0
pyyaml>=6.0,<7.0
tabulate>=0.9.0,<1.0             # jolis tableaux dans la console

# ─── Tests ──────────────────────────────────────────────────────────────
pytest>=8.0.0,<9.0
pytest-cov>=4.1.0,<6.0
